[video][android] set autoEnterPiP to false on video unmount
