+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoModule.kt
@@ -15,6 +15,8 @@ import expo.modules.kotlin.apifeatures.EitherType
 import expo.modules.kotlin.modules.Module
 import expo.modules.kotlin.modules.ModuleDefinition
 import expo.modules.kotlin.types.Either
+import expo.modules.video.VideoManager.isPictureInPictureSupported
+import expo.modules.video.VideoManager.runWithPiPMisconfigurationSoftHandling
 import expo.modules.video.enums.ContentFit
 import expo.modules.video.player.VideoPlayer
 import expo.modules.video.records.BufferOptions
@@ -36,7 +38,7 @@ class VideoModule : Module() {
     }
 
     Function("isPictureInPictureSupported") {
-      return@Function VideoView.isPictureInPictureSupported(appContext.throwingActivity)
+      return@Function isPictureInPictureSupported(appContext.throwingActivity)
     }
 
     View(VideoView::class) {
@@ -133,7 +135,7 @@ class VideoModule : Module() {
       }
 
       AsyncFunction("startPictureInPicture") { view: VideoView ->
-        view.runWithPiPMisconfigurationSoftHandling(true) {
+        runWithPiPMisconfigurationSoftHandling(true) {
           view.enterPictureInPicture()
         }
       }
@@ -143,7 +145,7 @@ class VideoModule : Module() {
       }
 
       OnViewDestroys {
-        VideoManager.unregisterVideoView(it)
+        VideoManager.unregisterVideoView(it, appContext)
       }
     }
